Split code snippet components into separate files
