+++ src/components/CodeSnippet/CodeSnippet.stories.mdx
@@ -1,7 +1,9 @@
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs/blocks";
 import { useState } from "react";
 
-import CodeSnippet, { CodeSnippetBlock, CodeSnippetDropdown, CodeSnippetBlockAppearance } from "./CodeSnippet";
+import CodeSnippet from "./CodeSnippet";
+import CodeSnippetBlock, { CodeSnippetBlockAppearance } from "./CodeSnippetBlock";
+import CodeSnippetDropdown from "./CodeSnippetDropdown";
 
 <Meta
   title="CodeSnippet"
