+++ src/components/CodeSnippet/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./CodeSnippet";
+export { CodeSnippetBlockAppearance } from "./CodeSnippetBlock";
